Unify Tasks with Task
