+++ src/db/scripts/exportTasks.js
@@ -1,4 +1,4 @@
-import Tasks from './Tasks.js';
+import Task from '../Task.js';
 import fs from 'node:fs';
 import path from 'path';
 import { fileURLToPath } from 'url';
@@ -9,7 +9,7 @@ const mariusUserID = "6uIcMMsBEkQ85OINCDADtrygzZx1"; //TESTING
 
 
 export async function exportTasks(user, fileName){
-    const data = await Tasks.getByDateRange(user, '2024-02-10', '2024-12-12');
+    const data = await Task.getByDateRange(user, '2024-02-10', '2024-12-12');
     fs.writeFile(path.join(__dirname, 'data', fileName + '.txt') , JSON.stringify(data), err => {
         if (err) {
           console.error(err);
